Compatible with TF2.2, code cleanup
**FEATURES**:
  - Compatibility with TF 2.2 (other versions still compatible, but no longer tested)

**BREAKING**:
  - `utils_225tf.py` removed
  - `utils_common.py` removed
  - `optimizers_tfpy.py` removed
  - `utils.py` code is now that of `utils_225tf.py`
  - `utils_common.py` merged with `utils.py`
