+++ keras_adamw/README.md
@@ -3,8 +3,8 @@
 ```python
 TensorFlow 1.14.0 + Keras 2.2.5 + 'keras'    >> optimizers_225.py   + utils.py
 TensorFlow 1.14.0 + Keras 2.2.5 + 'tf.keras' >> optimizers_225tf.py + utils_225tf.py
-TensorFlow 2.0.0  + Keras 2.3.0 + 'keras'    >> optimizers.py       + utils.py
-TensorFlow 2.0.0  + Keras 2.3.0 + 'tf.keras' >> optimizers_v2.py    + utils.py
+TensorFlow 2+     + Keras 2.3.0 + 'keras'    >> optimizers.py       + utils.py
+TensorFlow 2+     + Keras 2.3.0 + 'tf.keras' >> optimizers_v2.py    + utils.py
 ```
 
 [__init__.py](https://github.com/OverLordGoldDragon/keras-adamw/blob/master/keras_adamw/__init__.py) takes care of making the correct selection, but
